refactor: move comps into their own area
