+++ apps/playground/src/components/Comps/TAlert.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="my-2">
+    <h2>Alerts</h2>
+    <b-alert variant="primary" show>A simple primary alert—check it out!</b-alert>
+    <b-alert variant="warning" show dismissible>
+      <strong>Holy guacamole!</strong> You should check in on some of those fields below.
+    </b-alert>
+  </div>
+</template>
