+++ apps/playground/src/components/Comps/TBadge.vue
@@ -1,0 +1,56 @@
+<template>
+  <div class="my-2">
+    <h2>Badge</h2>
+    <div>
+      <h4>Simple</h4>
+      <div class="ms-4">
+        <h2>
+          Example heading
+          <b-badge>New</b-badge>
+        </h2>
+        <h3>
+          Example heading
+          <b-badge>New</b-badge>
+        </h3>
+        <h4>
+          Example heading
+          <b-badge>New</b-badge>
+        </h4>
+        <h5>
+          Example heading
+          <b-badge>New</b-badge>
+        </h5>
+        <h6>
+          Example heading
+          <b-badge>New</b-badge>
+        </h6>
+      </div>
+    </div>
+    <div class="mt-2">
+      <h4>Pill badges with color variant</h4>
+      <div class="ms-4">
+        <b-badge pill variant="primary">Primary</b-badge>
+        <b-badge pill variant="secondary">Secondary</b-badge>
+        <b-badge pill variant="success">Success</b-badge>
+        <b-badge pill variant="danger">Danger</b-badge>
+        <b-badge pill variant="warning">Warning</b-badge>
+        <b-badge pill variant="info">Info</b-badge>
+        <b-badge pill variant="light">Light</b-badge>
+        <b-badge pill variant="dark">Dark</b-badge>
+      </div>
+    </div>
+    <div class="mt-2">
+      <h4>Actionable badges</h4>
+      <div class="ms-4">
+        <b-badge href="#" variant="primary">Primary</b-badge>
+        <b-badge href="#" variant="secondary">Secondary</b-badge>
+        <b-badge href="#" variant="success">Success</b-badge>
+        <b-badge href="#" variant="danger">Danger</b-badge>
+        <b-badge href="#" variant="warning">Warning</b-badge>
+        <b-badge href="#" variant="info">Info</b-badge>
+        <b-badge href="#" variant="light">Light</b-badge>
+        <b-badge href="#" variant="dark">Dark</b-badge>
+      </div>
+    </div>
+  </div>
+</template>
