+++ apps/playground/src/components/Comps/TCard.vue
@@ -1,0 +1,83 @@
+<template>
+  <div class="my-2">
+    <h2 class="my-3">Cards</h2>
+    <b-row>
+      <b-col>
+        <b-card
+          no-body
+          style="max-width: 20rem"
+          img-src="https://placekitten.com/380/200"
+          img-alt="Image"
+          img-top
+        >
+          <template #header>
+            <h4 class="mb-0">Hello World</h4>
+          </template>
+
+          <b-card-body>
+            <b-card-title>Card Title</b-card-title>
+            <b-card-sub-title class="mb-2">Card Sub Title</b-card-sub-title>
+            <b-card-text>
+              Some quick example text to build on the card title and make up the bulk of the card's
+              content.
+            </b-card-text>
+          </b-card-body>
+
+          <b-list-group flush>
+            <b-list-group-item>Cras justo odio</b-list-group-item>
+            <b-list-group-item>Dapibus ac facilisis in</b-list-group-item>
+            <b-list-group-item>Vestibulum at eros</b-list-group-item>
+          </b-list-group>
+
+          <b-card-body>
+            <a href="#" class="card-link">Card link</a>
+            <a href="#" class="card-link">Another link</a>
+          </b-card-body>
+
+          <b-card-footer>This is a footer</b-card-footer>
+
+          <b-card-img src="https://placekitten.com/480/210" alt="Image" bottom />
+        </b-card>
+      </b-col>
+      <b-col>
+        <b-card
+          title="Card Title"
+          img-src="https://picsum.photos/600/300/?image=25"
+          img-alt="Image"
+          img-top
+          tag="article"
+          style="max-width: 20rem"
+          class="mb-2"
+        >
+          <b-card-text>
+            Some quick example text to build on the card title and make up the bulk of the card's
+            content.
+          </b-card-text>
+
+          <b-button href="#" variant="primary">Go somewhere</b-button>
+        </b-card>
+      </b-col>
+      <b-col cols="6">
+        <b-card no-body class="overflow-hidden" style="max-width: 540px">
+          <b-row class="g-0">
+            <b-col md="6">
+              <b-card-img
+                src="https://picsum.photos/400/400/?image=20"
+                alt="Image"
+                class="rounded-0"
+              />
+            </b-col>
+            <b-col md="6">
+              <b-card-body title="Horizontal Card">
+                <b-card-text>
+                  This is a wider card with supporting text as a natural lead-in to additional
+                  content. This content is a little bit longer.
+                </b-card-text>
+              </b-card-body>
+            </b-col>
+          </b-row>
+        </b-card>
+      </b-col>
+    </b-row>
+  </div>
+</template>
