+++ apps/playground/src/components/Comps/TFormFloatingLabel.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="my-2">
+    <h2>Form Floating Label</h2>
+    <div class="input-group mb-3">
+      <span id="basic-addon1" class="input-group-text">@</span>
+      <b-form-floating-label label="Email address">
+        <b-form-input placeholder=" " />
+      </b-form-floating-label>
+      <span class="input-group-text">.00</span>
+    </div>
+    <div>
+      <b-form-floating-label label="Email address">
+        <b-form-input placeholder=" " />
+      </b-form-floating-label>
+    </div>
+  </div>
+</template>
