+++ apps/playground/src/components/Comps/TListGroup.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="my-2">
+    <h2>List Group</h2>
+    <div>
+      <b-list-group>
+        <b-list-group-item>Cras justo odio</b-list-group-item>
+        <b-list-group-item>Dapibus ac facilisis in</b-list-group-item>
+        <b-list-group-item>Morbi leo risus</b-list-group-item>
+        <b-list-group-item>Porta ac consectetur ac</b-list-group-item>
+        <b-list-group-item>Vestibulum at eros</b-list-group-item>
+      </b-list-group>
+    </div>
+  </div>
+</template>
