+++ apps/playground/src/components/Comps/TNav.vue
@@ -1,0 +1,36 @@
+<template>
+  <div class="my-2">
+    <h2>Nav</h2>
+    <h4 class="my-3">Nav with a dropdown</h4>
+    <div>
+      <b-nav pills>
+        <b-nav-item active>Active</b-nav-item>
+        <b-nav-item>Link</b-nav-item>
+        <b-nav-item-dropdown
+          id="my-nav-dropdown"
+          text="Dropdown"
+          toggle-class="nav-link-custom"
+          right
+        >
+          <b-dropdown-item>One</b-dropdown-item>
+          <b-dropdown-item>Two</b-dropdown-item>
+          <b-dropdown-divider></b-dropdown-divider>
+          <b-dropdown-item>Three</b-dropdown-item>
+        </b-nav-item-dropdown>
+      </b-nav>
+    </div>
+    <h4 class="m-2">Nav with a dropdown and a custom Button Icon</h4>
+    <div>
+      <b-nav pills>
+        <b-nav-item active>Active</b-nav-item>
+        <b-nav-item>Link</b-nav-item>
+        <b-nav-item-dropdown text="Custom Button Icon" no-caret variant="link">
+          <template #button-content>
+            <img src="./assets/logo.png" style="height: 1em" />
+          </template>
+          <b-dropdown-item href="#">Action</b-dropdown-item>
+        </b-nav-item-dropdown>
+      </b-nav>
+    </div>
+  </div>
+</template>
