+++ apps/playground/src/components/Comps/TSkeletonTable.vue
@@ -1,0 +1,3 @@
+<template>
+  <b-skeleton-table :rows="4" :columns="3" :table-props="{bordered: true, striped: true}" />
+</template>
