+++ apps/playground/src/components/Comps/TTable.vue
@@ -1,0 +1,103 @@
+<template>
+  <div class="my-2">
+    <h2>Table</h2>
+    <div>
+      <h4 class="my-3">Table without field definitions</h4>
+      <b-table responsive caption="List of users" :items="items" striped hover foot-clone>
+        <template #cell(first_name)="data">
+          <a href="#">{{ data.value }}</a>
+        </template>
+      </b-table>
+    </div>
+    <div>
+      <h4 class="my-3">Table with string field definitions</h4>
+      <b-table
+        responsive
+        caption="List of users"
+        :items="items"
+        :fields="stringTableDefinitions"
+        striped
+        hover
+        foot-clone
+      >
+        <template #cell(first_name)="data">
+          <a href="#">{{ data.value }}</a>
+        </template>
+      </b-table>
+    </div>
+    <div>
+      <h4 class="my-3">Table with object field definitions</h4>
+
+      <b-table
+        responsive
+        caption="List of users"
+        :items="items"
+        :fields="objectTableDefinitions"
+        striped
+        hover
+        foot-clone
+      >
+        <template #cell(first_name)="data">
+          <a href="#">{{ data.value }}</a>
+        </template>
+      </b-table>
+    </div>
+    <div>
+      <h4 class="my-3">Table: Adding additional rows to the header</h4>
+
+      <b-table
+        responsive
+        caption="List of users"
+        :items="items"
+        :fields="objectTableDefinitions"
+        striped
+        hover
+        foot-clone
+        dark
+      >
+        <template #thead-top>
+          <tr class="my">
+            <th colspan="2">
+              <span class="sr-only">List of users</span>
+            </th>
+          </tr>
+        </template>
+        <template #thead-sub="{key, label}">
+          <tr class="my">
+            <th variant="danger" />
+            <th variant="danger">
+              <b-form-select :placeholder="label" :options="[label, key]" label-field="label" />
+            </th>
+          </tr>
+        </template>
+        <template #cell(first_name)="data">
+          <a href="#">{{ data.value }}</a>
+        </template>
+      </b-table>
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+import {ref, Ref} from 'vue'
+import {TableField, TableItem} from 'bootstrap-vue-3'
+
+const stringTableDefinitions = ref(['last_name', 'first_name', 'age'])
+const objectTableDefinitions: Ref<Array<TableField>> = ref([
+  {key: 'last_name', label: 'Family name'},
+  {key: 'first_name', label: 'Given name'},
+])
+const items: Array<TableItem> = [
+  {age: 40, first_name: 'Dickerson', last_name: 'Macdonald'},
+  {age: 21, first_name: 'Larsen', last_name: 'Shaw'},
+  {age: 89, first_name: 'Geneva', last_name: 'Wilson'},
+  {
+    age: 89,
+    first_name: 'Michael',
+    last_name: 'Jordan',
+    _rowVariant: 'primary',
+    _cellVariants: {last_name: 'danger'},
+  },
+  {age: 38, first_name: 'Jami', last_name: 'Carney'},
+]
+</script>
